test(fixtures): expand and reorganize tests
* add tests for dynamically generated example/testmodel fixtures
* move test markers to separate module
* checkout testmodel and example repos in CI
