+++ .github/workflows/ci.yml
@@ -3,11 +3,11 @@ on:
   push:
     branches:
       - main
-      - develop
+      - develop*
   pull_request:
     branches:
       - main
-      - develop
+      - develop*
 jobs:
   lint:
     name: Lint
@@ -49,7 +49,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ ubuntu-latest, macos-latest, windows-latest ]
-        python: [ 3.7, 3.8, 3.9, "3.10" ]
+        python: [ 3.8, 3.9, "3.10" ]
         exclude:
           # avoid shutil.copytree infinite recursion bug
           # https://github.com/python/cpython/pull/17098
@@ -66,6 +66,24 @@ jobs:
         with:
           repository: MODFLOW-USGS/modflow6
           path: modflow6
+      
+      - name: Checkout modflow6 examples
+        uses: actions/checkout@v3
+        with:
+          repository: MODFLOW-USGS/modflow6-examples
+          path: modflow6-examples
+      
+      - name: Checkout modflow6 test models
+        uses: actions/checkout@v3
+        with:
+          repository: MODFLOW-USGS/modflow6-testmodels
+          path: modflow6-testmodels
+      
+      - name: Checkout modflow6 large test models
+        uses: actions/checkout@v3
+        with:
+          repository: MODFLOW-USGS/modflow6-largetestmodels
+          path: modflow6-largetestmodels
 
       - name: Setup Python
         uses: actions/setup-python@v4
@@ -77,8 +95,18 @@ jobs:
       - name: Install Python packages
         working-directory: modflow-devtools
         run: |
-          python3 -m pip install .
-          python3 -m pip install ".[test]"
+          pip install .
+          pip install ".[test]"
+      
+      - name: Install extra Python packages
+        working-directory: modflow6-examples/etc
+        run: |
+          pip install -r requirements.pip.txt
+          pip install -r requirements.usgs.txt
+
+      - name: Build modflow6 example models
+        working-directory: modflow6-examples/etc
+        run: python ci_build_files.py
 
       - name: Run tests
         working-directory: modflow-devtools
